Server_3.4.0 updateLog: 1、增加FAQ功能 2、修复部分手动出入库的bug

+++ uw_server/src/main/java/com/jimi/uw_server/service/TaskService.java
@@ -34,7 +34,6 @@ import com.jimi.uw_server.constant.TaskState;
 import com.jimi.uw_server.constant.TaskType;
 import com.jimi.uw_server.exception.OperationException;
 import com.jimi.uw_server.lock.Lock;
-import com.jimi.uw_server.model.ErrorLog;
 import com.jimi.uw_server.model.ExternalWhLog;
 import com.jimi.uw_server.model.FormerSupplier;
 import com.jimi.uw_server.model.GoodsLocation;
@@ -361,10 +360,11 @@ public class TaskService {
 			} else if (state == TaskState.CANCELED) { // 对于已作废过的任务，禁止作废
 				throw new OperationException("该任务已作废！");
 			} else {
-
-				// 当前仓口存在未放入料盒的入库物料
-				if ((task.getType().equals(TaskType.IN) || task.getType().equals(TaskType.SEND_BACK)) && !InputMaterialRedisDAO.getScanStatus(task.getWindow()).equals(-1)) {
-					throw new OperationException("禁止作废，请先将之前的物料正确入库再进行作废任务操作");
+				if (Integer.valueOf(PropKit.use("properties.ini").get("input_open")) == 1) {
+					// 当前仓口存在未放入料盒的入库物料
+					if ((task.getType().equals(TaskType.IN) || task.getType().equals(TaskType.SEND_BACK)) && !InputMaterialRedisDAO.getScanStatus(task.getWindow()).equals(-1)) {
+						throw new OperationException("禁止作废，请先将之前的物料正确入库再进行作废任务操作");
+					}
 				}
 				boolean untiedWindowflag = true;
 				// 判断任务是否处于进行中状态，若是，则把相关的任务条目从til中剔除;若存在已分配的任务条目，则不解绑任务仓口
@@ -905,12 +905,12 @@ public class TaskService {
 					windowId = item.getWindowId();
 				}
 			}
-
-			// 当前仓口存在未放入料盒的入库物料
-			if (!InputMaterialRedisDAO.getScanStatus(windowId).equals(-1)) {
-				throw new OperationException("扫码错误，请先将之前的物料正确入库再进行扫码操作");
+			if (Integer.valueOf(PropKit.use("properties.ini").get("input_open")) == 1) {
+				// 当前仓口存在未放入料盒的入库物料
+				if (!InputMaterialRedisDAO.getScanStatus(windowId).equals(-1)) {
+					throw new OperationException("扫码错误，请先将之前的物料正确入库再进行扫码操作");
+				}
 			}
-
 			Task task = Task.dao.findById(packingListItem.getTaskId());
 			Integer reelNum = 0;
 			if (materialType.getRadius().equals(7)) {
@@ -978,24 +978,26 @@ public class TaskService {
 			taskLog.setTime(new Date());
 			taskLog.setDestination(task.getDestination());
 			taskLog.save();
-			if (!material.getCol().equals(-1) && !material.getRow().equals(-1)) {
-				int positionNo = material.getCol() * 20 + material.getRow();
-				String result = MyInputHelper.getInstance().switchLight(windowId, positionNo, true);
-				if (result.contains("succeed")) {
-					InputMaterialRedisDAO.setScanStatus(windowId, positionNo);
-				} else {
-					taskLog.delete();
-					if (material.getIsRepeated() != null && material.getIsRepeated() ) {
-						material.setRemainderQuantity(0);
-						material.setCol(-1);
-						material.setRow(-1);
-						material.setIsInBox(false);
-						material.setIsRepeated(true);
-						material.update();
+			if (Integer.valueOf(PropKit.use("properties.ini").get("input_open")) == 1) {
+				if (!material.getCol().equals(-1) && !material.getRow().equals(-1)) {
+					int positionNo = material.getCol() * 20 + material.getRow();
+					String result = MyInputHelper.getInstance().switchLight(windowId, positionNo, true);
+					if (result.contains("succeed")) {
+						InputMaterialRedisDAO.setScanStatus(windowId, positionNo);
 					} else {
-						material.delete();
+						taskLog.delete();
+						if (material.getIsRepeated() != null && material.getIsRepeated() ) {
+							material.setRemainderQuantity(0);
+							material.setCol(-1);
+							material.setRow(-1);
+							material.setIsInBox(false);
+							material.setIsRepeated(true);
+							material.update();
+						} else {
+							material.delete();
+						}
+						throw new OperationException("入库失败，开启投料辅助器LED灯失败|Error|LED OPERN FAILED|" + result);
 					}
-					throw new OperationException("入库失败，开启投料辅助器LED灯失败|Error|LED OPERN FAILED|" + result);
 				}
 			}
 			material.setIsRepeated(false).update();
@@ -1350,8 +1352,10 @@ public class TaskService {
 				throw new OperationException("盘点任务UW仓盘点阶段已结束，无法指定或更改仓口！");
 			}
 		}
-		if ((task.getType().equals(TaskType.IN) || task.getType().equals(TaskType.SEND_BACK)) && !InputMaterialRedisDAO.getScanStatus(task.getWindow()).equals(-1)) {
-			throw new OperationException("无法指定或更改仓口，请先将之前的物料正确入库再进行更改仓口操作");
+		if (Integer.valueOf(PropKit.use("properties.ini").get("input_open")) == 1) {
+			if ((task.getType().equals(TaskType.IN) || task.getType().equals(TaskType.SEND_BACK)) && !InputMaterialRedisDAO.getScanStatus(task.getWindow()).equals(-1)) {
+				throw new OperationException("无法指定或更改仓口，请先将之前的物料正确入库再进行更改仓口操作");
+			}
 		}
 		List<Integer> windowIdList = new ArrayList<>();
 		if (windowIds != null && !windowIds.trim().equals("")) {
